+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleDataproc.scala
@@ -6,7 +6,7 @@ import cats.mtl.ApplicativeAsk
 import com.google.api.gax.core.FixedCredentialsProvider
 import com.google.auth.oauth2.ServiceAccountCredentials
 import com.google.cloud.dataproc.v1._
-import io.chrisdavenport.log4cats.Logger
+import io.chrisdavenport.log4cats.StructuredLogger
 import org.broadinstitute.dsde.workbench.RetryConfig
 import org.broadinstitute.dsde.workbench.model.TraceId
 import org.broadinstitute.dsde.workbench.model.google.GcsBucketName
@@ -30,14 +30,14 @@ trait GoogleDataproc[F[_]] {
 }
 
 object GoogleDataproc {
-  def fromCredentialPath[F[_]: Logger: Async: Timer: ContextShift](pathToCredential: String, blocker: Blocker, blockerBound: Semaphore[F], retryConfig: RetryConfig = GoogleDataprocInterpreter.defaultRetryConfig): Resource[F, GoogleDataproc[F]] = for {
+  def fromCredentialPath[F[_]: StructuredLogger: Async: Timer: ContextShift](pathToCredential: String, blocker: Blocker, blockerBound: Semaphore[F], retryConfig: RetryConfig = GoogleDataprocInterpreter.defaultRetryConfig): Resource[F, GoogleDataproc[F]] = for {
     credentialFile <- org.broadinstitute.dsde.workbench.util2.readFile(pathToCredential)
     client <- fromServiceAccountCrendential(ServiceAccountCredentials.fromStream(credentialFile), blocker, blockerBound, retryConfig)
   } yield client
 
-  def fromServiceAccountCrendential[F[_]: Logger: Async: Timer: ContextShift](serviceAccountCredentials: ServiceAccountCredentials, blocker: Blocker, blockerBound: Semaphore[F], retryConfig: RetryConfig = GoogleTopicAdminInterpreter.defaultRetryConfig): Resource[F, GoogleDataproc[F]] = {
+  def fromServiceAccountCrendential[F[_]: StructuredLogger: Async: Timer: ContextShift](serviceAccountCredentials: ServiceAccountCredentials, blocker: Blocker, blockerBound: Semaphore[F], retryConfig: RetryConfig = GoogleTopicAdminInterpreter.defaultRetryConfig): Resource[F, GoogleDataproc[F]] = {
     val settings = ClusterControllerSettings.newBuilder()
-      .setCredentialsProvider(FixedCredentialsProvider.create(serviceAccountCredentials))
+      .setCredentialsProvider(FixedCredentialsProvider.create(serviceAccountCredentchials))
       .build()
 
     for {
@@ -45,7 +45,7 @@ object GoogleDataproc {
     } yield new GoogleDataprocInterpreter[F](client, retryConfig, blocker, blockerBound)
   }
 
-  def fromApplicationDefault[F[_]: ContextShift: Timer: Async: Logger](blocker: Blocker, blockerBound: Semaphore[F], retryConfig: RetryConfig = GoogleDataprocInterpreter.defaultRetryConfig): Resource[F, GoogleDataproc[F]] = for {
+  def fromApplicationDefault[F[_]: ContextShift: Timer: Async: StructuredLogger](blocker: Blocker, blockerBound: Semaphore[F], retryConfig: RetryConfig = GoogleDataprocInterpreter.defaultRetryConfig): Resource[F, GoogleDataproc[F]] = for {
     db <- Resource.make(
       Sync[F].delay(
         ClusterControllerClient.create()