+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GooglePublisher.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.dsde.workbench.google2
 
 import cats.effect.{Async, ContextShift, Resource, Timer}
 import fs2.Pipe
-import io.chrisdavenport.log4cats.Logger
+import io.chrisdavenport.log4cats.StructuredLogger
 import io.circe.Encoder
 
 trait GooglePublisher[F[_]] {
@@ -14,12 +14,17 @@ trait GooglePublisher[F[_]] {
   /**
     * Watch out message size quota and limitations https://cloud.google.com/pubsub/quotas
     */
+  def tracedPublish[MessageType: Encoder]: Pipe[F, Message[MessageType], Unit]
+
+  /**
+    * Watch out message size quota and limitations https://cloud.google.com/pubsub/quotas
+    */
   def publishString: Pipe[F, String, Unit]
 
 }
 
 object GooglePublisher {
-  def resource[F[_]: Async: Timer: ContextShift: Logger, MessageType: Encoder](config: PublisherConfig): Resource[F, GooglePublisher[F]] = for {
+  def resource[F[_]: Async: Timer: ContextShift: StructuredLogger, MessageType: Encoder](config: PublisherConfig): Resource[F, GooglePublisher[F]] = for {
     publisher <- GooglePublisherInterpreter.publisher(config)
   } yield GooglePublisherInterpreter(publisher, config.retryConfig)
 }